+++ boosie_ui/util/scheduler.h
@@ -15,7 +15,7 @@ public:
 	static void start();
 	static void stop();
 
-	static void schedule(menu_callback_t callback, uint64_t arg = 0, uint64_t interval = 0, thread thread = render);
+	static void schedule(menu_callback_t callback, uint32_t arg = 0, uint64_t interval = 0, thread thread = render);
 	static void once(menu_callback_t callback, thread thread = render);
 	static void delay(menu_callback_t callback, uint64_t delay, thread thread = render);
 
@@ -23,7 +23,7 @@ public:
 	{
 		scheduler::thread thread;
 		menu_callback_t callback;
-		uint64_t argument;
+		uint32_t argument;
 		uint64_t interval;
 		uint64_t last_call_time;
 		bool is_temp;
